request constructors moved to separate layer
